Move quotes into App
