+++ src/apiCalls.js
@@ -1,9 +1,12 @@
-const apiCalls = {
-  fetchData(url){ 
-    return fetch(url)
-      .then(response => response.json())
+// const apiCalls = {
+//   fetchData(url) {
+//     return fetch(url, {
+//       mode: 'no-cors',
+//       headers: {
+//         'Access-Control-Allow-Origin'
+//       }
+//     }).then(response => response.json())
+//   }
+// }
 
-  }
-}
-
-export default apiCalls
+// export default apiCalls
