+++ src/components/App/App.js
@@ -6,10 +6,46 @@ import { Route, Routes, Link, useNavigate } from 'react-router-dom'
 import GameScreen from '../GameScreen/GameScreen'
 import EndGame from '../End Game/EndGame'
 import shakeSpeareQuotes from '../../shakeSpearData'
+import taylorSwiftQuotes from '../../shakeSpearData'
 import Error from '../Error/Error'
 import HomeScreen from '../HomseScreen/HomeScreen'
 
 function App() {
+
+  const shakeSpeareQuotes = [
+    'Love all, trust a few, do wrong to none.',
+    'Love looks not with the eyes, but with the mind.',
+    'Be not afraid of greatness. Some are born great, some achieve greatness, and others have greatness thrust upon them.',
+    'There is nothing either good or bad, but thinking makes it so.',
+    'Hell is empty and all the devils are here.',
+    'We know what we are, but not what we may be.',
+    'You speak an infinite deal of nothing.',
+    'The course of true love never did run smooth.',
+    "Don't waste your love on somebody, who doesn't value it.",
+    'Expectation is the root of all heartache.',
+    'Listen to many, speak to a few.',
+    'One may smile, and smile, and be a villain.',
+    "What's done cannot be undone.",
+    'All the worlds a stage, and all the men and women merely players. They have their exits and their entrances; and one man in his time plays many parts.',
+    'All that glisters is not gold.',
+    'I like this place and could willingly waste my time in it.'
+  ]
+
+  const taylorSwiftQuotes = [
+    'This is a new year.A new beginning.And things will change.',
+    'You took a swing, I took it hard. And down here from the ground I see who you are',
+    'No matter what happens in life, be good to people. Being good to people is a wonderful legacy to leave behind.',
+    'I never want to change so much that people cant recognize me.',
+    'Im intimidated by the fear of being average.',
+    'Just be yourself, there is no one better.',
+    'Words can break someone into a million pieces, but they can also put them back together. I hope you use yours for good, because the only words youll regret more than the ones left unsaid are the ones you use to intentionally hurt someone.',
+    'You can write a book on how to ruin someones perfect day.',
+    'giving up doesnt always mean your weak sometimes your just strong enough to let go',
+    "So don't you worry your pretty little mind because people throw rocks at things that shine",
+    "The way you walk, way you talk, way you say my name; it's beautiful, wonderful, don't you ever change.",
+    "The only one who's got enough of me to break my heart.",
+    "I've wanted one thing for me whole life and I'm not going to be that girl who wants one thing her whole life then gets it and complains."
+  ]
   const [timer, setTimer] = useState(30)
   const [isTimeSelected, setIsTimeSelected] = useState(false)
   const [isEndGame, setIsEndGame] = useState(false)
@@ -18,36 +54,38 @@ function App() {
   const [isItSwiftOrSpeare, setIsitSwiftOrSpeare] = useState([])
   const [displayedQuotes, setDisplayedQuotes] = useState([])
   const [userGuess, setUserGuess] = useState([])
-  const [taylorSwift, setTaylorSwift] = useState([])
-  const [taylorSwiftUsableQuotes, setTaylorSwiftUsableQuotes] = useState([])
+  const [taylorSwift, setTaylorSwift] = useState(taylorSwiftQuotes)
+  const [taylorSwiftUsableQuotes, setTaylorSwiftUsableQuotes] = useState([...taylorSwift])
   const [speareUsableQuotes, setSpeareUsableQuotes] = useState([
     ...shakeSpeareQuotes
   ])
   const [shakeSpeare, setShakeSpeare] = useState(shakeSpeareQuotes)
   const [ranOutOfQuotes, setRanOutOfQuotes] = useState(false)
 
-  useEffect(() => {
-    apiCalls
-      .fetchData('https://taylorswiftapi.herokuapp.com/get-all?album=fearless')
-      .then(data => {
-        setTaylorSwift(data)
-      })
-  }, [])
+  
 
-  useEffect(() => {
-    if (!isGameStarted) {
-      let cleanedQuotes = []
-      taylorSwift.forEach(tquote => {
-        if (tquote.quote.includes('/')) {
-          let cleaned = tquote.quote.replace(/\//g, '')
-          cleanedQuotes.push(cleaned)
-        } else {
-          cleanedQuotes.push(tquote.quote)
-        }
-      })
-      setTaylorSwiftUsableQuotes([...cleanedQuotes])
-    }
-  }, [taylorSwift, isGameStarted])
+  // useEffect(() => {
+  //   apiCalls
+  //     .fetchData('https://taylorswiftapi.herokuapp.com/get-all?album=fearless')
+  //     .then(data => {
+  //       setTaylorSwift(data)
+  //     })
+  // }, [])
+
+  // useEffect(() => {
+  //   if (!isGameStarted) {
+  //     let cleanedQuotes = []
+  //     taylorSwift.forEach(tquote => {
+  //       if (tquote.quote.includes('/')) {
+  //         let cleaned = tquote.quote.replace(/\//g, '')
+  //         cleanedQuotes.push(cleaned)
+  //       } else {
+  //         cleanedQuotes.push(tquote.quote)
+  //       }
+  //     })
+  //     setTaylorSwiftUsableQuotes([...cleanedQuotes])
+  //   }
+  // }, [taylorSwift, isGameStarted])
 
   useEffect(() => {
     if (!isGameStarted) {
